i18n load textdomain and language files
Find language files in gatherpress/languages when missing in wp-content/languages/plugins/

The translation files will be in wp-content/languages/plugins/ once the plugin on the repository and translated in translate.wordpress.org.
Until that, we need to load from /languages folder and load the textdomain.
See https://developer.wordpress.org/plugins/internationalization/how-to-internationalize-your-plugin/#plugins-on-wordpress-org