Fixes issue with timing the filter
Sets the initial duration value to null instead of using the getDateTimeOffset function.
This function is moved to the getDuration selector. This selector is run in the Duration component which renders after any theme filters have been registered. We do need to check if the value is false when the user selects the value to custom pick the end time.